move validateAbsolutePaths to validation.ts
