+++ components/webui/client/src/pages/IngestPage/Compress/PathsInputFormItem.tsx
@@ -1,32 +1,10 @@
-import {Value} from "@sinclair/typebox/value";
-import {AbsolutePathSchema} from "@webui/common/schemas/compression";
-import {Nullable} from "@webui/common/utility-types";
 import {
     Form,
     Input,
 } from "antd";
 
+import {validateAbsolutePaths} from "./validation";
 
-/**
- * Validates that all non-empty lines in the input are absolute paths using AbsolutePathSchema.
- *
- * @param value The multiline string input containing paths.
- * @return An error message if validation fails, or null if validation passes.
- */
-const validateAbsolutePaths = (value: string): Nullable<string> => {
-    const lines = value.split("\n");
-    for (const [index, line] of lines.entries()) {
-        const trimmedLine = line.trim();
-        if (0 === trimmedLine.length) {
-            continue;
-        }
-        if (false === Value.Check(AbsolutePathSchema, trimmedLine)) {
-            return `Line ${index + 1}: Path must be absolute (start with "/").`;
-        }
-    }
-
-    return null;
-};
 
 /**
  * Renders a form item for inputting file paths for compression job submission.
